+++ source-code/cli/src/utilities/getGitRemotes.ts
@@ -6,13 +6,13 @@ import fs from "node:fs"
  *
  * @returns The git origin url or undefined if it could not be found.
  */
-export async function getGitOrigin() {
+export async function getGitRemotes() {
 	try {
 		const remotes = await raw.listRemotes({
 			fs,
 			dir: await raw.findRoot({ fs, filepath: process.cwd() }),
 		})
-		return remotes.find((remote) => remote.remote === "origin")?.url
+		return remotes
 	} catch (e) {
 		return undefined
 	}